refactor(internal/ent): 修改用户ID类型为int并调整相关代码
修复了用户ID类型的不一致问题，并在多个文件中进行了相应的调整。

+++ internal/ent/runtime.go
@@ -44,5 +44,5 @@ func init() {
 	// userDescID is the schema descriptor for id field.
 	userDescID := userFields[0].Descriptor()
 	// user.IDValidator is a validator for the "id" field. It is called by the builders before save.
-	user.IDValidator = userDescID.Validators[0].(func(uint64) error)
+	user.IDValidator = userDescID.Validators[0].(func(int) error)
 }
